Update Function.java
The initial step of making the below scenario :

On the post req, i am gonna mention the firstName and lastName :
Ex:-
{
    "firstName": "John",
    "lastName": "Jayathunga"
}

Then if the first name and last name fulfill the requirement as it is, the output should be like the photo (Tute03) i provided now. it means the firstname and lastname i provided on the post, the status code with a "0" and a message mentioned as "Success". And if there are any kind of a miss or unrelated input or input in a wrong input type or etc, the output should consist of the firstname and lastname provided on the post, the status code with a "-1" and a message mentioned as "Unsuccessful".